+++ lib/manageiq/providers/red_hat_virtualization.rb
